refactor package sructure, add new handlers
