--- api/oapi-models-codegen.yaml
@@ -1,4 +1,0 @@
-package: models
-output: models/models.gen.go
-generate:
-  models: true
