+++ internal/api/handlers.go
@@ -1,0 +1,41 @@
+package api
+
+import "net/http"
+
+type HandlerConfig struct {
+	// Add configuration fields as needed
+}
+
+type Handlers struct {
+	Config interface{}
+}
+
+func NewHandlers(config HandlerConfig) *Handlers {
+	return &Handlers{
+		Config: config,
+	}
+}
+
+func (h *Handlers) CreateCase(w http.ResponseWriter, r *http.Request) {
+	// Register your routes here
+}
+
+func (h *Handlers) CreateEvent(w http.ResponseWriter, r *http.Request) {
+	// Register your routes here
+}
+
+func (h *Handlers) ListCases(w http.ResponseWriter, r *http.Request) {
+	// Register your routes here
+}
+
+func (h *Handlers) ListEvents(w http.ResponseWriter, r *http.Request) {
+	// Register your routes here
+}
+
+func (h *Handlers) ReadCase(w http.ResponseWriter, r *http.Request) {
+	// Register your routes here
+}
+
+func (h *Handlers) ReadEvent(w http.ResponseWriter, r *http.Request) {
+	// Register your routes here
+}
