+++ internal/api/router.go
@@ -1,0 +1,9 @@
+package api
+
+import (
+	"net/http"
+)
+
+func NewRouter() *http.ServeMux {
+	return http.NewServeMux()
+}
